# Conflicts:
#	ChronoLoop/ChronoLoop/Actions/CCButtonPress.h
#	ChronoLoop/ChronoLoop/Actions/TeleportAction.hpp

+++ ChronoLoop/ChronoLoop/Particles/ParticleSystem.cpp
@@ -157,7 +157,7 @@ namespace Epoch
 
 	void ParticleSystem::Clear()
 	{
-		for (int i = mPEmitters.size() - 1; i >= 0; i--)
+		for (unsigned int i = (unsigned int)mPEmitters.size() - 1; i >= 0; i--)
 		{
 			if (DoesExist(mPEmitters[i]))
 			{
